+++ src/client/GraphQLClient/client.test.ts
@@ -15,6 +15,7 @@ import {
   REPOS_QUERY_STRING,
   TEAM_MEMBERS_QUERY_STRING,
   USERS_QUERY_STRING,
+  MAX_REQUESTS_NUM,
 } from './queries';
 import { parseTimePropertyValue } from '@jupiterone/integration-sdk-core';
 
@@ -42,17 +43,34 @@ async function getAccess() {
   return { token, tokenExpires, appClient };
 }
 
-const pageLimit = 2;
 async function getClient() {
   const { token, tokenExpires, appClient } = await getAccess();
   return new GitHubGraphQLClient(
     token,
     tokenExpires,
-    resourceMetadataMap(pageLimit, pageLimit),
+    resourceMetadataMap(),
     createMockIntegrationLogger(),
     appClient,
   );
 }
+
+//reduce page limit size so we can test pagination
+const pullRequestsQueryString = PULL_REQUESTS_QUERY_STRING.replace(
+  `first: ${MAX_REQUESTS_NUM}`,
+  'first: 2',
+);
+const reposQueryString = REPOS_QUERY_STRING.replace(
+  `first: ${MAX_REQUESTS_NUM}`,
+  'first: 2',
+);
+const usersQueryString = USERS_QUERY_STRING.replace(
+  `first: ${MAX_REQUESTS_NUM}`,
+  'first: 2',
+);
+const teamMembersQueryString = TEAM_MEMBERS_QUERY_STRING.replace(
+  `first: ${MAX_REQUESTS_NUM}`,
+  'first: 2',
+);
 
 describe('pull request resources', () => {
   let p: Recording; //p for polly
@@ -67,11 +85,10 @@ describe('pull request resources', () => {
       name: 'GitHubGraphQLClient.fetchPullRequests.noExtraResources',
     });
     const client = await getClient();
-
     const query = 'is:pr repo:JupiterOne/graph-whitehat is:open';
     const pullRequests: PullRequest[] = [];
     const response = await client.iteratePullRequests(
-      PULL_REQUESTS_QUERY_STRING,
+      pullRequestsQueryString,
       query,
       [],
       (pr) => {
@@ -107,11 +124,10 @@ describe('pull request resources', () => {
       name: 'GitHubGraphQLClient.fetchPullRequests.singlePage',
     });
     const client = await getClient();
-
     const query = 'is:pr repo:Kei-Institute/Test-repo is:open';
     const pullRequests: PullRequest[] = [];
     const response = await client.iteratePullRequests(
-      PULL_REQUESTS_QUERY_STRING,
+      pullRequestsQueryString,
       query,
       [GithubResource.Commits, GithubResource.Reviews, GithubResource.Labels],
       (pr) => {
@@ -129,12 +145,11 @@ describe('pull request resources', () => {
       name: 'GitHubGraphQLClient.fetchPullRequests.innerPagination',
     });
     const client = await getClient();
-
     const query =
       'is:pr repo:JupiterOne/graph-whitehat is:closed updated:<=2019-04-01 ';
     const pullRequests: PullRequest[] = [];
     const data = await client.iteratePullRequests(
-      PULL_REQUESTS_QUERY_STRING,
+      pullRequestsQueryString,
       query,
       [GithubResource.Commits, GithubResource.Reviews, GithubResource.Labels],
       (pr) => {
@@ -182,7 +197,7 @@ describe('organization resources', () => {
     const client = await getClient();
 
     const data = await client.fetchFromSingle(
-      REPOS_QUERY_STRING,
+      reposQueryString,
       GithubResource.Organization,
       [GithubResource.Repositories],
       { login: 'Kei-Institute' },
@@ -209,7 +224,7 @@ describe('organization resources', () => {
     const client = await getClient();
 
     const data = await client.fetchFromSingle(
-      USERS_QUERY_STRING,
+      usersQueryString,
       GithubResource.Organization,
       [GithubResource.OrganizationMembers],
       { login: 'Kei-Institute' },
@@ -236,7 +251,7 @@ describe('organization resources', () => {
     const client = await getClient();
 
     const data = await client.fetchFromSingle(
-      TEAM_MEMBERS_QUERY_STRING,
+      teamMembersQueryString,
       GithubResource.Organization,
       [GithubResource.TeamMembers],
       { login: 'Kei-Institute' },